add possibility to pass in envelop
