+++ src/index.ts
@@ -1,3 +1,4 @@
 export { default as default } from './helix-flare'
+export { default as helixFlareEnvelop } from './envelop'
 export { createExecutor } from './createExecutor'
 export { createSubscription, STOP } from './createSubscription'
